requirements.txt
openai
speechRecognition
python-dotenv
pipwin
PyAudio

